// src/glsl/utilityFunctions.d.ts
export declare const utilityFunctions: {
    _luminance: {
        type: string;
        glsl: string;
    };
    _noise: {
        type: string;
        glsl: string;
    };
    _rgbToHsv: {
        type: string;
        glsl: string;
    };
    _hsvToRgb: {
        type: string;
        glsl: string;
    };
    _rotate: {
        type: string;
        glsl: string;
    };
};